validatePopupValues
