+++ public_modules/const/locators.mjs
@@ -7,3 +7,13 @@ export const editTitleBtn = document.getElementById("sloikTitleEdit");
 export const editDescriptionBtn = document.getElementById("sloikDescriptionEdit");
 
 export const addMoneyBtn = document.getElementById("addMoneyBtn");
+
+export const popup = document.getElementById("popupContainer");
+export const popupDescription = document.getElementById("popupDescription");
+export const titleInput = document.getElementById("inputField1");
+export const descriptionInput = document.getElementById("inputField2");
+export const goalSumInput = document.getElementById("inputField3");
+export const inputError1 = document.getElementById("inputField1Error");
+export const inputError2 = document.getElementById("inputField2Error");
+export const inputError3 = document.getElementById("inputField3Error");
+export const editPopupBtn = document.getElementById("editBtn");
